autoencoder_model/dataset.py
import os
from torch.utils.data import Dataset
from PIL import Image

class ClothDataset(Dataset):
    def __init__(self, img_dir, transform=None):
        self.img_dir = img_dir
        self.transform = transform
        self.paths = [os.path.join(img_dir, fname) for fname in os.listdir(img_dir)]

    def __len__(self):
        return len(self.paths)

    def __getitem__(self, idx):
        img_path = self.paths[idx]
        image = Image.open(img_path).convert('RGB')
        if self.transform:
            image = self.transform(image)
        return image, img_path